Fix sorting of paged results

+++ core/src/main/java/com/gentics/mesh/core/data/page/Page.java
@@ -44,7 +44,7 @@ public interface Page {
 	 * 
 	 * @return
 	 */
-	int getTotalElements();
+	long getTotalElements();
 
 	/**
 	 * Return the page size.
